picovoice/requirements.txt
paho-mqtt
pvporcupine
pvrhino
pyaudio
numpy
